fix(bumble): address fourth round of review
* Move generic helpers out of `transport/bumble/__init__.py`:
  - `bumble_device()` + host/HCI constants → new `bumble/device.py`
  - `pair()`, `pair_device()`, `encrypt_using_bond()` + pairing constants →
    `bumble/pairing.py` (was just delegates + result types)
* `SMPBumbleTransport`: split `clear_bond(address)` and `clear_bonds()`
  (no `None`-coded meaning), inline `bonds` local in `bonded_devices`,
  drop the impl-detail line from its docstring.
* `_notifications` / `_disconnected_event`: `Final` annotations.
* `_next_chunk`: walrus-style `match (chunk := await ...)`.
* `_on_security_request`: pattern-match on state and log at the right level
  (info for borrowed, warn for Disconnected/Connecting, debug for Connected).
* Add `borrowed_connection()` async context manager wrapping
  `use_connection()` + `disconnect()`.
* Trim narrative comments where the code is the explanation.
